Create HelloWorld in GoogleMap
It's the code of Java in our first Project. With the help of this code we can create a new application, extend from PApplet. We will create a canvas and draw 2 limited maps in it. One of them is a view of UCSD's campus, and another is a view of my hometown Munich. Actually it's not my hometown but a city that I'm studying in.